refactor(desktop): extract memory ipc
